+++ test/test_models/common.py
@@ -12,15 +12,20 @@ from gptorch.util import torch_dtype
 torch.set_default_dtype(torch_dtype)
 
 
-def gaussian_predictions(model: GPModel, x_test: torch.Tensor, 
-        expected_mu: np.ndarray, expected_s: np.ndarray):
+def gaussian_predictions(
+    model: GPModel,
+    x_test: torch.Tensor,
+    expected_mu: np.ndarray,
+    expected_s: np.ndarray,
+):
     """
     Every GP model with a gaussian likelihood needs the same set of tests run on
     its ._predict() method.
     """
 
     # Predictions without full covariance
-    mu_diag, s_diag = model._predict(x_test, diag=True)
+    pf_diag = model._predict(x_test, diag=True)
+    mu_diag, s_diag = pf_diag.loc, pf_diag.variance
 
     assert isinstance(mu_diag, torch.Tensor)
     assert isinstance(s_diag, torch.Tensor)
@@ -29,22 +34,41 @@ def gaussian_predictions(model: GPModel, x_test: torch.Tensor,
     assert mu_diag.shape[1] == model.Y.shape[1]
     assert all([ss == ms for ss, ms in zip(mu_diag.shape, s_diag.shape)])
 
-    assert all([a == pytest.approx(e) for a, e in zip(
-        mu_diag.detach().numpy().flatten(), expected_mu.flatten())])
-    assert all([a == pytest.approx(e) for a, e in zip(
-        s_diag.detach().numpy().flatten(), expected_s.diagonal().flatten())])
+    assert all(
+        [
+            a == pytest.approx(e)
+            for a, e in zip(mu_diag.detach().numpy().flatten(), expected_mu.flatten())
+        ]
+    )
+    assert all(
+        [
+            a == pytest.approx(e)
+            for a, e in zip(
+                s_diag.detach().numpy().flatten(), expected_s.diagonal().flatten()
+            )
+        ]
+    )
 
     # Predictions with full covariance
-    mu_full, s_full = model._predict(x_test, diag=False)
+    pf_full = model._predict(x_test, diag=False)
+    mu_full, s_full = pf_full.loc.T, pf_full.covariance_matrix
 
     assert isinstance(mu_full, torch.Tensor)
     assert isinstance(s_full, torch.Tensor)
 
     assert mu_full.shape[0] == x_test.shape[0]
     assert mu_full.shape[1] == model.Y.shape[1]
-    assert all([ss == x_test.shape[0] for ss in s_full.shape])
-    
-    assert all([a == pytest.approx(e) for a, e in zip(
-        mu_full.detach().numpy().flatten(), expected_mu.flatten())])
-    assert all([a == pytest.approx(e) for a, e in zip(
-        s_full.detach().numpy().flatten(), expected_s.flatten())])
+    assert all([ss == x_test.shape[0] for ss in s_full.shape[-2:]])
+
+    assert all(
+        [
+            a == pytest.approx(e)
+            for a, e in zip(mu_full.detach().numpy().flatten(), expected_mu.flatten())
+        ]
+    )
+    assert all(
+        [
+            a == pytest.approx(e)
+            for a, e in zip(s_full.detach().numpy().flatten(), expected_s.flatten())
+        ]
+    )